// src/hooks/types.ts
export interface FeedbackFormData {
  subject: string
  url: string
  comment: string
}